creating abstract client class and interface; implementing invdvidual client class

+++ Accounts/IAccount.cs
@@ -8,7 +8,7 @@ namespace Bank.Accounts
 {
     internal interface IAccount
     {
-        void MakeATransaction(int accountNumber, double amount, double services, string description = "", bool statement = true);
+        void MakeATransaction(long accountNumber, double amount, double services, string description = "", bool statement = true);
         void FundsWithdrawal(double amount, bool statement = true);
         void FundsDeposit(double amount, bool statement = true);
     }
